requirements.txt
# Google Cloud - Gemini models
google-generativeai>=0.8.0

# AWS Bedrock - Claude and Nova2 models
boto3>=1.35.0

# Config & env
pyyaml>=6.0
python-dotenv>=1.0.0
